--- lib/ReactViews/Credits/Credits.props.ts
@@ -1,5 +1,0 @@
-import { ICredit } from "./Credit.type";
-
-export interface ICreditsProps {
-  credits?: ICredit[];
-}
