--- lib/ReactViews/Credits/DataAttribution/DataAttibutionModal.props.ts
@@ -1,4 +1,0 @@
-export interface IDataAttributionModalProps {
-  closeModal: () => void;
-  attributions?: string[];
-}
