--- lib/ReactViews/Credits/MapCredits/MapCreditLogo.props.ts
@@ -1,5 +1,0 @@
-import GlobeOrMap from "../../../Models/GlobeOrMap";
-
-export interface IMapCreditLogoProps {
-  currentViewer: GlobeOrMap;
-}
