--- lib/ReactViews/Credits/MapCredits/MapCredits.props.ts
@@ -1,8 +1,0 @@
-import { ICredit } from "../Credit.type";
-import GlobeOrMap from "../../../Models/GlobeOrMap";
-
-export interface IMapCreditsProps {
-  hideTerriaLogo: boolean;
-  credits?: ICredit[];
-  currentViewer: GlobeOrMap;
-}
