+++ lib/ReactViews/Credits/index.ts
@@ -1,1 +1,2 @@
+export { ICredit } from "./Credit.type";
 export { MapCredits } from "./MapCredits/MapCredits";
